Convertet from Typscript to Javascript, WiFi Thermostat driver and device.
